requirements.txt
streamlit==1.25.0
plotly==5.9.0
pandas==1.5.3
numpy==1.23.5
folium==0.14.0
matplotlib==3.7.0
matplotlib-inline==0.1.6
haversine==2.8.0
streamlit-folium==0.13.0
Pillow==9.4.0
